fix: esm resolution error with pkg.type -> `module`
Currently, when running `npx opentype.js/ot` or installing `opentype.js` and running `ot`, an ESModule resolution error is thrown. 

This is due to Node assuming the `.js` file extensions only contain CommonJS code, when in fact they are in ES6 syntax.

Adding `"type": "module"` to the package.json forces all `.js` files to be interpreted as ES6 modules instead.